requirements.txt
python-dotenv
llama-index
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai
llama-parse
llama-index-vector-stores-chroma
chromadb
llama-index-callbacks-arize-phoenix
llama-index-tools-tavily-research
nest-asyncio
llama-index-utils-workflow
sqlalchemy
streamlit
unstructured[all-docs]
langchain-openai
langchain-community
langchain-core
llama-index-program-openai
openinference-instrumentation-langchain
arize-phoenix-otel